* Add Disconnection of player on timeout. * WIP directional movement component

+++ Engine/Managers/SceneManager.cpp
@@ -58,7 +58,7 @@ void SceneManager::notifyDisplayFrame()
 	glClearColor(0.0, 0.0, 0.0, 1.0);
 	
 
-	for (int i = 0; i < gameObjectManager.size(); i++)
+	for (size_t i = 0; i < gameObjectManager.size(); i++)
 	{
 		assert(!mainCamera.expired());
 		if (!mainCamera.expired())
